feat: 完成post頁面所有api串接，更新動態路由
